Implement ProjectileBase, a class which will be the foundation for projectiles
