+++ Source/BrackeysGameJam2020/Private/ProjectileBase.cpp
@@ -1,0 +1,37 @@
+
+
+
+#include "ProjectileBase.h"
+#include "GameFramework/ProjectileMovementComponent.h"
+#include "Components/StaticMeshComponent.h"
+
+// Sets default values
+AProjectileBase::AProjectileBase()
+{
+ 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
+	PrimaryActorTick.bCanEverTick = true;
+
+	//Create dodgeball static mesh component and make it the root component
+	StaticMesh = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("Projectile Mesh"));
+	SetRootComponent(StaticMesh);
+
+	//Create Projectile Movement Component which will handle the projectile movement (Speed, bouncing, etc)
+	ProjectileMovement = CreateDefaultSubobject<UProjectileMovementComponent>(TEXT("Projectile Movement"));
+}
+
+// Called when the game starts or when spawned
+void AProjectileBase::BeginPlay()
+{
+	Super::BeginPlay();
+
+	//Bind OnActorHit to my own OnHitActor() function so when OnActorHit is triggered, OnHitActor() is called
+	StaticMesh->OnComponentBeginOverlap.AddDynamic(this, &ADodgeBall::OnOverlapComponent);
+}
+
+// Called every frame
+void AProjectileBase::Tick(float DeltaTime)
+{
+	Super::Tick(DeltaTime);
+
+}
+
