+++ Source/BrackeysGameJam2020/Public/DodgeBall.h
@@ -4,10 +4,9 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/Actor.h"
+#include "ProjectileBase.h"
 #include "DodgeBall.generated.h"
 
-class UProjectileMovementComponent;
-class UStaticMeshComponent;
 class APlayerBase;
 
 //Enumerator for dodge ball state machine
@@ -20,7 +19,7 @@ enum Projectile
 };
 
 UCLASS()
-class BRACKEYSGAMEJAM2020_API ADodgeBall : public AActor
+class BRACKEYSGAMEJAM2020_API ADodgeBall : public AProjectileBase
 {
 	GENERATED_BODY()
 	
@@ -34,14 +33,6 @@ public:
 
 	// Called every frame
 	virtual void Tick(float DeltaTime) override;
-
-	//Projectile Movement Component
-	UPROPERTY(VisibleAnywhere, Category = "Components")
-	UProjectileMovementComponent* ProjectileMovement;
-
-	//Projectile Mesh Component
-	UPROPERTY(VisibleAnywhere, Category = "Components")
-	UStaticMeshComponent* StaticMesh;
 
 	//Called to modify whether the ball is returning to the player
 	UFUNCTION(BlueprintCallable, Category = "Dodgeball")
@@ -59,12 +50,9 @@ public:
 	UPROPERTY(EditDefaultsOnly, Category = "Dodgeball")
 	float BallReturnDelay = 2.f;
 
-	UPROPERTY(EditDefaultsOnly, Category = "Dodgeball")
-	float DamageAmount = 10.f;
-
 	//Called when the projectile hits something
 	UFUNCTION()
-	void OnOverlapComponent(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult);
+	void OnOverlapComponent(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult) override;
 
 	//Accessor to get bCanCurve
 	bool GetCanCurve() const { return bCanCurve; }
