+++ Source/BrackeysGameJam2020/Public/ProjectileBase.h
@@ -1,0 +1,44 @@
+
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/Actor.h"
+#include "ProjectileBase.generated.h"
+
+class UProjectileMovementComponent;
+class UStaticMeshComponent;
+
+UCLASS()
+class BRACKEYSGAMEJAM2020_API AProjectileBase : public AActor
+{
+	GENERATED_BODY()
+	
+public:	
+	// Sets default values for this actor's properties
+	AProjectileBase();
+
+protected:
+	// Called when the game starts or when spawned
+	virtual void BeginPlay() override;
+
+public:	
+	// Called every frame
+	virtual void Tick(float DeltaTime) override;
+
+	//Projectile Movement Component
+	UPROPERTY(VisibleAnywhere, Category = "Components")
+	UProjectileMovementComponent* ProjectileMovement;
+
+	//Projectile Mesh Component
+	UPROPERTY(VisibleAnywhere, Category = "Components")
+	UStaticMeshComponent* StaticMesh;
+
+	UPROPERTY(EditDefaultsOnly, Category = "Dodgeball")
+	float DamageAmount = 10.f;
+
+	//Called when the projectile hits something
+	UFUNCTION()
+	virtual void OnOverlapComponent(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult);
+
+};
